Add tox.ini
This is required to be able to build and release the project using
the standard infra tools.  Since there currently aren't any unit
test or doc jobs defined for this project, only a very basic
tox.ini is being added.  It can be filled out later as more jobs
are needed.

Change-Id: I121405d4ec82a4d8252098319fce0a9f5b5bfb4f