Use 'output' function instead of 'echo' for statistics log
In bash the only sensible way of returning a string from functions
or scripts is via stdout.
dib-run-parts not only 'returned' the output of the called scripts but
also some statistics / performance output.
Other parts of the script already used an 'output' function so
the statistics output was adapted to also use this kind of function.

In addition three simple test cases are implemented to check
the basic functions of the script.

As a result the log-output of the dib-run-parts is somewhat cleaner.
It also makes debugging of the called scripts easier.
This change is the precondition for simplification of functions like
diskimage-builder's eval_run_d.

Change-Id: Ie2f0332359a72ca7ab03f94b7573e073df714a97
Signed-off-by: Andreas Florath <andreas@florath.net>