Source environment files in a subshell
By sourcing the environment files directly into the running process,
dib-run-parts can pollute its own environment.  For example, if an
environment script enables tracing (set -x) then that enables tracing
for dib-run-parts; particularly annoying as it then spews a bunch of
stuff relating to the profiling.

Move to sourcing the environment files in a subshell with the scripts
to be executed to keep things separate.

Change-Id: I1e39822f218dc0322e2490a770f3dc867a55802c